+++ geoLocalize/locationIQ.go
@@ -1,12 +1,45 @@
 package geolocalize
 
 import (
+	"encoding/json"
 	"fmt"
+
+	"github.com/dompaw/RumAir/airStations"
 )
 
 var locationiqBaseApiURL string = "https://locationiq.org/v1/reverse.php?key=e281731b38bb74"
 
-func getReverseGeocodedCitiesLocationIQ(radius int, lat float64, lon float64) (bytesRead []byte, err error) {
+func LocalizeStationsLocIQ(stations map[string]*airStations.AirStation) (result map[string]*LocalizedAirStation, err error) {
+	result = map[string]*LocalizedAirStation{}
+
+	for id, station := range stations {
+		if station.LatitudeSensor != "" && station.LongitudeSensor != "" {
+			if localizedStation, err := LocalizeStationLocIQ(station); err == nil {
+				result[id] = localizedStation
+			}
+		}
+	}
+	return
+}
+
+func LocalizeStationLocIQ(station *airStations.AirStation) (result *LocalizedAirStation, err error) {
+	result = &LocalizedAirStation{Station: station}
+	if result.Lat, result.Lon, err = getStationCoordinates(station); err == nil && result.Lat != 0 && result.Lon != 0 {
+		result.CitiesNearby, err = getCitiesNearbyLocIQ(result.Lat, result.Lon)
+		return
+	}
+	return result, fmt.Errorf("Can't localize station %v lat and lon. \n", station.ID)
+}
+
+func getReverseGeocodedCitiesLocIQ(lat float64, lon float64) (bytesRead []byte, err error) {
+	//https://locationiq.org/v1/reverse.php?key=e281731b38bb74&lat=54.5987&lon=18.26669336&format=json
+	locationIqURL := locationiqBaseApiURL + fmt.Sprintf("&lat=%f&lon=%f&format=json", lat, lon)
+	fmt.Printf("ENDPOINT : %s. \n", locationIqURL)
+
+	return doAPIGet(locationIqURL)
+}
+
+func getCitiesNearbyLocIQ(lat float64, lon float64) (citiesNearby []string, err error) {
 	type LocationIQReverseGeoResponse struct {
 		Address struct {
 			CityDistrict string `json:"city_district"`
@@ -26,18 +59,21 @@ func getReverseGeocodedCitiesLocationIQ(radius int, lat float64, lon float64) (b
 		Lon         string `json:"lon"`
 	}
 
-	//https://locationiq.org/v1/reverse.php?key=e281731b38bb74&lat=54.5987&lon=18.26669336&format=json
-	locationIqURL := locationiqBaseApiURL + fmt.Sprintf("?callback=RumAir&radius=%v&latitude=%f&longitude=%f", radius, lat, lon)
-	fmt.Printf("ENDPOINT : %s. \n", locationIqURL)
-
-	return doAPIGet(locationIqURL)
-}
-
-/*func getCitiesNearby(lat float64, lon float64) (citiesNearby []string, err error) {
-	radius := 0
-	var reverseGeocodingStringedResponce string
 	var bytesRead []byte
+	var result *LocationIQReverseGeoResponse
+	bytesRead, err = getReverseGeocodedCitiesLocIQ(lat, lon)
 
-	for until := true; until; until = (len(reverseGeocodingStringedResponce)) < 5 {
+	if err != nil {
+		fmt.Printf("Error during ReadAll bytesRead: %s err: %v. \n", bytesRead, err)
 	}
-}*/
+
+	if len(bytesRead) > 0 {
+		err = json.Unmarshal(bytesRead, &result)
+		if err != nil {
+			fmt.Printf("Error during deserializing occured. Data from `../table=Measurement&v=2`. Error is : %v", err)
+			return
+		}
+		citiesNearby = append(citiesNearby, fmt.Sprintf("%s, %s", result.Address.Town, result.Address.Suburb))
+	}
+	return
+}
